requirements.txt
fastapi
paddlepaddle
paddleocr
pillow
pydantic